rename: 폴더구조 및 네이밍 변경
