+++ src/App.tsx
@@ -2,16 +2,16 @@ import { ThemeProvider } from 'styled-components';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { theme } from './styles';
 import GlobalStyle from './styles/GlobalStyle';
-import RouteChangeTracker from './components/RouteChangeTracker.tsx';
+import useRouteTracker from './hooks/useRouteTracker.tsx';
 import Main from './pages/Main.tsx';
-import UserLogIn from './pages/UserLogIn.tsx';
-import AdminLogIn from './pages/AdminLogIn.tsx';
+import UserLogIn from './pages/LogIn/indexUser.tsx';
+import AdminLogIn from './pages/LogIn/indexAdmin.tsx';
 
 function App() {
+  useRouteTracker();
   return (
     <ThemeProvider theme={theme}>
       <GlobalStyle />
-      <RouteChangeTracker />
       <BrowserRouter>
         <Routes>
           <Route path="/" element={<Main />} />
