+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "types": ["node"],
+    "types": ["node", "vite/client"],
     "target": "ES2020",
     "useDefineForClassFields": true,
     "lib": ["ES2020", "DOM", "DOM.Iterable"],
@@ -18,5 +18,5 @@
     "noFallthroughCasesInSwitch": true,
     "allowSyntheticDefaultImports": true
   },
-  "include": ["src", "vite.config.dev.ts", "vite.config.prod.ts", "vite.config.ts"]
+  "include": ["src/**/*", "vite.config.dev.ts", "vite.config.prod.ts", "vite.config.ts"]
 }
